% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineCoverages.R
\name{combineCoverages}
\alias{combineCoverages}
\title{Combine observed and predicted coverages}
\usage{
combineCoverages(junctionCounts, predictedCoverages, txQuants)
}
\arguments{
\item{junctionCounts}{data.frame with observed junction coverages}

\item{predictedCoverages}{data.frame with predicted junction coverages}

\item{txQuants}{data.frame with estimated transcript abundances}
}
\value{

}
\description{
Description...
}
\examples{
jcov <- read.delim(junctioncovSTAR, header = FALSE, as.is = TRUE)
colnames(jcov) <- c("seqnames", "start", "end", "strand", "motif", "annot",
                    "uniqreads", "mmreads", "maxoverhang")
jcov <- jcov \%>\% dplyr::mutate(strand = replace(strand, strand == 1, "+")) \%>\%
  dplyr::mutate(strand = replace(strand, strand == 2, "-")) \%>\%
  dplyr::select(seqnames, start, end, strand, uniqreads, mmreads)

}
\references{
Soneson C, Love MI, Patro R, Hussain S, Malhotra D, Robinson MD: A junction coverage compatibility score to quantify the reliability of transcript abundance estimates and annotation catalogs. bioRxiv doi:10.1101/378539 (2018)
}
\author{
Charlotte Soneson
}
