% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAlpineBiasModel.R
\name{fitAlpineBiasModel}
\alias{fitAlpineBiasModel}
\title{Fit fragment bias model with alpine}
\usage{
fitAlpineBiasModel(gtf, bam, organism, genome, genomeVersion, version,
  minLength = 600, maxLength = 7000, minCount = 500, maxCount = 10000,
  subsample = TRUE, nbrSubsample = 200, verbose = FALSE)
}
\arguments{
\item{gtf}{gtf file with genomic features. Preferably in Ensembl format}

\item{bam}{bam file with read alignments to the genome}

\item{organism}{The organism}

\item{genome}{A BSgenome object}

\item{genomeVersion}{Genome version}

\item{version}{version}

\item{minLength, maxLength}{Minimum and maximum length of single-isoform genes
used to fit fragment bias model}

\item{minCount, maxCount}{Minimum and maximum read coverage of single-isoform
genes used to fit fragment bias model}

\item{subsample}{Whether to subsample the set of single-isoform genes
satisfying the \code{minLength}, \code{maxLength}, \code{minCount} and
\code{maxCount} criteria before fitting fragment bias model}

\item{nbrSubsample}{If \code{subsample} is \code{TRUE}, how many genes to
subsample}

\item{verbose}{Whether to print progress messages}
}
\description{
Fit fragment bias model with alpine
}
\author{
Charlotte Soneson, Michael I Love
}
