% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateJCCScores.R
\name{gthr}
\alias{gthr}
\title{Weight function}
\usage{
gthr(omega, thr)
}
\arguments{
\item{omega}{Fraction of uniquely mapped reads among those spanning the
junction.}

\item{thr}{Threshold value. If \code{omega} is above this value, the junction
contributes to the JCC score of the gene. If \code{omega} is below this
value, the junction doesn't contribute.}
}
\value{
Either 1 or 0, representing the weight for the junction.
}
\description{
Function to determine the weight for a junction in the calculation of the JCC
scores and the scaled coverages. This function is used by
\code{junctionScore} and \code{scaledCoverage}, and does not have to be
called by the user.
}
\author{
Charlotte Soneson
}
