% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateJCCScores.R
\name{scaledCoverage}
\alias{scaledCoverage}
\title{Help function for calculation of scaled junction coverages}
\usage{
scaledCoverage(uniqreads, mmreads, predcovs, g, beta = 1, ...)
}
\arguments{
\item{uniqreads}{Vector with the number of uniquely mapping junction-spanning
reads for each junction in the gene.}

\item{mmreads}{Vector with the number of multimapping junction-spanning reads
for each junction in the gene.}

\item{predcovs}{Vector with the predicted junction coverages for each
junction in the gene.}

\item{g}{Junction weight function that takes as input the fraction of
uniquely mapping reads for the junction, and returns a weight for the
junction contribution to the score.}

\item{beta}{Scaling exponent. If beta=1, predicted coverages are scaled to
have the same (weighted) sum as the observed coverages. If beta=0, no
scaling is done.}

\item{...}{Additional parameters passed to \code{g}.}
}
\value{
A vector of scaled junction coverages
}
\description{
Function to calculate the scaled junction coverages for a given gene
}
\author{
Charlotte Soneson
}
