% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleTxCoverages.R
\name{scaleTxCoverages}
\alias{scaleTxCoverages}
\title{Scale transcript coverages by abundance estimate}
\usage{
scaleTxCoverages(txCoverageProfiles, txQuants, tx2Gene, strandSpecific,
  verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
Scale transcript coverages by abundance estimate
}
\author{
Charlotte Soneson
}
