% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateJCCScores.R
\name{calculateJCCScores}
\alias{calculateJCCScores}
\title{Calculate scaled coverages and JCC scores}
\usage{
calculateJCCScores(junctionCounts, txQuantsGene, mmfracthreshold = 0.25)
}
\arguments{
\item{mmfracthreshold}{}
}
\description{
Calculate scaled coverages and JCC scores
}
\author{
Charlotte Soneson
}
