% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineCoverages.R
\name{combineCoverages}
\alias{combineCoverages}
\title{Combine observed and predicted coverages}
\usage{
combineCoverages(junctionCounts, predictedCoverages, txQuants)
}
\arguments{
\item{txQuants}{}
}
\description{
Combine observed and predicted coverages
}
\examples{
jcov <- read.delim(junctioncovSTAR, header = FALSE, as.is = TRUE)
colnames(jcov) <- c("seqnames", "start", "end", "strand", "motif", "annot",
                    "uniqreads", "mmreads", "maxoverhang")
jcov <- jcov \%>\% dplyr::mutate(strand = replace(strand, strand == 1, "+")) \%>\%
  dplyr::mutate(strand = replace(strand, strand == 2, "-")) \%>\%
  dplyr::select(seqnames, start, end, strand, uniqreads, mmreads)
}
\author{
Charlotte Soneson
}
