% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAlpineBiasModel.R
\name{fitAlpineBiasModel}
\alias{fitAlpineBiasModel}
\title{Fit fragment bias model with alpine}
\usage{
fitAlpineBiasModel(gtf, bam, organism, genome, genomeVersion, version,
  minLength = 600, maxLength = 7000, minCount = 500, maxCount = 10000,
  subsample = TRUE, nbrSubsample = 200, minSize = NULL, maxSize = NULL,
  verbose = FALSE)
}
\arguments{
\item{gtf}{gtf file with genomic features. Preferably in Ensembl format}

\item{bam}{bam file with read alignments to the genome}

\item{organism}{The organism}

\item{genome}{A BSgenome object}

\item{genomeVersion}{Genome version}

\item{version}{version}

\item{minLength, maxLength}{Minimum and maximum length of single-isoform genes
used to fit fragment bias model}

\item{minCount, maxCount}{Minimum and maximum read coverage of single-isoform
genes used to fit fragment bias model}

\item{subsample}{Whether to subsample the set of single-isoform genes
satisfying the \code{minLength}, \code{maxLength}, \code{minCount} and
\code{maxCount} criteria before fitting fragment bias model}

\item{nbrSubsample}{If \code{subsample} is \code{TRUE}, how many genes to
subsample}

\item{minSize, maxSize}{Smallest and largest fragment size to consider. If any
of these is null, it is estimated from the data.}

\item{verbose}{Logical, whether to print progress messages.}
}
\value{

}
\description{
Description
}
\examples{

}
\references{
Soneson C, Love MI, Patro R, Hussain S, Malhotra D, Robinson MD: A junction coverage compatibility score to quantify the reliability of transcript abundance estimates and annotation catalogs. bioRxiv doi:10.1101/378539 (2018)
}
\author{
Charlotte Soneson, Michael I Love
}
