% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictTxCoverage.R
\name{predictTxCoverage}
\alias{predictTxCoverage}
\title{Predict coverage profiles of transcripts}
\usage{
predictTxCoverage(biasModel, exonsByTx, bam, genes, nCores, tx2Gene, genome,
  verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
Predict coverage profiles of transcripts
}
\author{
Charlotte Soneson, Michael I Love
}
